create test page for visualisation
